# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Todo Application API",
    "description": "A comprehensive Todo List application with categories, tags, and advanced filtering capabilities. This API provides full CRUD operations for managing todos, categories, and tags with extensive logging.",
    "contact": {
      "name": "Metrifuge",
      "email": "support@metrifuge.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
    "/api/todos": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get all todos",
        "description": "Retrieve all todos from the database",
        "operationId": "getAllTodos",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Todo"],
        "summary": "Create new todo",
        "description": "Create a new todo item",
        "operationId": "createTodo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todo by ID",
        "description": "Retrieve a specific todo by its ID",
        "operationId": "getTodoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the todo to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "put": {
        "tags": ["Todo"],
        "summary": "Update todo",
        "description": "Update an existing todo",
        "operationId": "updateTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the todo to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "delete": {
        "tags": ["Todo"],
        "summary": "Delete todo",
        "description": "Delete a todo by its ID",
        "operationId": "deleteTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the todo to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Todo deleted successfully"
          },
          "404": {
            "description": "Todo not found"
          }
        }
      }
    },
    "/api/todos/completed/{completed}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos by completion status",
        "description": "Filter todos by completed or incomplete status",
        "operationId": "getTodosByCompleted",
        "parameters": [
          {
            "name": "completed",
            "in": "path",
            "description": "Completion status",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/priority/{priority}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos by priority",
        "description": "Filter todos by priority level",
        "operationId": "getTodosByPriority",
        "parameters": [
          {
            "name": "priority",
            "in": "path",
            "description": "Priority level",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/status/{status}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos by status",
        "description": "Filter todos by workflow status",
        "operationId": "getTodosByStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Workflow status",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["TODO", "IN_PROGRESS", "BLOCKED", "REVIEW", "DONE"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/category/{categoryId}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos by category",
        "description": "Filter todos by category ID",
        "operationId": "getTodosByCategoryId",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/tag/{tagId}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos by tag",
        "description": "Filter todos by tag ID",
        "operationId": "getTodosByTagId",
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "description": "Tag ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/search": {
      "get": {
        "tags": ["Todo"],
        "summary": "Search todos",
        "description": "Search todos by keyword in title or description",
        "operationId": "searchTodos",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "Search keyword",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/assigned/{assignedTo}": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos by assignee",
        "description": "Filter todos by assigned user",
        "operationId": "getTodosByAssignedTo",
        "parameters": [
          {
            "name": "assignedTo",
            "in": "path",
            "description": "Username of assignee",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/due": {
      "get": {
        "tags": ["Todo"],
        "summary": "Get todos due in date range",
        "description": "Filter todos by due date range",
        "operationId": "getTodosDueBetween",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "Start date (ISO 8601 format)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "End date (ISO 8601 format)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "completed": {
            "type": "boolean",
            "default": false
          },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
            "default": "MEDIUM"
          },
          "status": {
            "type": "string",
            "enum": ["TODO", "IN_PROGRESS", "BLOCKED", "REVIEW", "DONE"],
            "default": "TODO"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "estimatedHours": {
            "type": "integer"
          },
          "actualHours": {
            "type": "integer"
          },
          "assignedTo": {
            "type": "string"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "Tag": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      }
    }
  }
}